requirements.txt
numpy==1.19.5
streamlit==0.85.0
fastai==2.4
torch==1.9.0+cu111
Pillow==8.3.1
